Move cleanup related functions to a separate module
Now:
- The cleanup functions can be read and used by any other version of the application
- The Pre2014 cleanup functions are different and stay with the transform_pre2014.py file
